requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmudict==1.1.1
coverage==7.8.2
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-model-utils==5.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jiter==0.10.0
kombu==5.5.4
lxml==6.0.0
model-bakery==1.20.4
openai==1.99.9
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
textstat==0.7.8
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.23.0
